Export parse id to separate file to fix build
